Complete Simulation 2 JSON & Refactor Escalation checkWarState()
Simulation 2 is now complete. Centurion and I have fixed the seg fault for loadEscaltion. In cases where the escaltion is the same between rounds, the same pointer was being returned instead of new objects and when the newWarState object was pointing to said object it threw an error as the previous warStage pointer was deleted. Was not fun.